feat: allow editing of search bar placeholder text in config editor
